move explain cmd/options to separate file
