+++ src/cmap/wire_protocol/query.ts
@@ -7,7 +7,6 @@ import { Document, pluckBSONSerializeOptions } from '../../bson';
 import type { Server } from '../../sdam/server';
 import type { ReadPreferenceLike } from '../../read_preference';
 import type { FindOptions } from '../../operations/find';
-import { Explain } from '../../explain';
 
 /** @internal */
 export interface QueryOptions extends CommandOptions {
@@ -63,7 +62,7 @@ export function query(
 }
 
 function prepareFindCommand(server: Server, ns: string, cmd: Document) {
-  const findCmd: Document = {
+  let findCmd: Document = {
     find: collectionNamespace(ns)
   };
 
@@ -147,9 +146,12 @@ function prepareFindCommand(server: Server, ns: string, cmd: Document) {
   if (cmd.collation) findCmd.collation = cmd.collation;
   if (cmd.readConcern) findCmd.readConcern = cmd.readConcern;
 
-  // TODO: Quick fix to make tests pass; will be updated during NODE-2853
-  if (cmd.explain !== undefined) {
-    findCmd.explain = Explain.fromOptions({ explain: cmd.explain });
+  // If we have explain, we need to rewrite the find command
+  // to wrap it in the explain command
+  if (cmd.explain) {
+    findCmd = {
+      explain: findCmd
+    };
   }
 
   return findCmd;
