+++ src/explain.ts
@@ -1,6 +1,5 @@
-import type { Callback, Document } from '.';
-import { MongoError } from './error';
-import { CommandOperation, CommandOperationOptions, OperationParent } from './operations/command';
+import type { Document } from '.';
+import type { ExplainOptions } from './operations/explainable_command';
 import type { Server } from './sdam/server';
 import { maxWireVersion } from './utils';
 
@@ -9,45 +8,6 @@ const SUPPORTS_EXPLAIN_WITH_UPDATE = 3;
 const SUPPORTS_EXPLAIN_WITH_DISTINCT = 3.2;
 const SUPPORTS_EXPLAIN_WITH_FIND_AND_MODIFY = 3.2;
 const SUPPORTS_EXPLAIN_WITH_MAP_REDUCE = 4.4;
-
-/** @internal */
-export abstract class ExplainableCommand<
-  T extends ExplainOptions = ExplainOptions,
-  TResult = Document
-> extends CommandOperation<T, TResult> {
-  explain?: Explain;
-
-  constructor(parent?: OperationParent, options?: T) {
-    super(parent, options);
-
-    if (!Explain.explainOptionsValid(options)) {
-      throw new MongoError(`explain must be one of ${Object.keys(Verbosity)} or a boolean`);
-    }
-
-    this.explain = Explain.fromOptions(options);
-  }
-
-  get canRetryWrite(): boolean {
-    return this.explain === undefined;
-  }
-
-  executeCommand(server: Server, cmd: Document, callback: Callback): void {
-    if (this.explain) {
-      if (!Explain.explainSupportedOnCmd(server, cmd)) {
-        callback(new MongoError(`server ${server.name} does not support explain on this command`));
-        return;
-      }
-
-      cmd.explain = this.explain;
-    }
-    super.executeCommand(server, cmd, callback);
-  }
-}
-
-/** @public */
-export interface ExplainOptions extends CommandOperationOptions {
-  explain?: VerbosityLike;
-}
 
 /** @public */
 export enum Verbosity {
@@ -81,7 +41,7 @@ export class Explain {
     return new Explain(options.explain);
   }
 
-  static explainOptionsValid(options?: ExplainOptions): boolean {
+  static valid(options?: ExplainOptions): boolean {
     if (options == null || options.explain === undefined) {
       return true;
     }
@@ -117,21 +77,4 @@ export class Explain {
       (cmd.mapReduce && wireVersion >= SUPPORTS_EXPLAIN_WITH_MAP_REDUCE)
     );
   }
-}
-
-/**
- * Applies an explain to a given command.
- * @internal
- *
- * @param command - the command on which to apply the read concern
- * @param options - the options containing the explain verbosity
- */
-export function decorateWithExplain(command: Document, explain: Explain): Document {
-  // A command being explained may not have an explain field directly on it
-  if (command.explain !== undefined) {
-    delete command.explain;
-  }
-
-  command = { explain: command, verbosity: explain.verbosity };
-  return command;
 }
